+++ .env
@@ -1,13 +1,8 @@
 PORT=5000
 
-MONGODB_URI=mongodb+srv://farhanrasta:<pingsuit2>@project-game-be.59eb7ne.mongodb.net/?retryWrites=true&w=majority
+MONGODB_URL=mongodb+srv://farhanrasta:pingsuit2@project-game-be.59eb7ne.mongodb.net/?retryWrites=true&w=majority
 JWT_SECRET=9d76f4eb9a55ee24c7b8ecbe352d42d6785b8e2f9747c3c48a5b6d8cb88d2fd5
 
 PASSWORD = pingsuit2
 
-
-//MONGODB_URI=mongodb+srv://saputramfirrizqi:8EuRSwsUsZo4BY49@project-game-be.fburbbj.mongodb.net/?retryWrites=true&w=majority
-//JWT_SECRET=9d76f4eb9a55ee24c7b8ecbe352d42d6785b8e2f9747c3c48a5b6d8cb88d2fd5
-
-//PASSWORD = 8EuRSwsUsZo4BY49
